fix: restore conservative advisor injection logic, move virtual registry test to internal/tool
- Revert isVirtualServerToolInjectable: no advisor source config → return
  false (not injectable), matching original behavior before the refactor
  introduced !foundAdvisorSource which inadvertently made unconfigured
  advisor injectable
- Move virtual_registry_test.go from internal/mcp/runtime/ to
  internal/tool/ where VirtualToolRegistry is now defined; update test
  to use package-local types directly
- Fix advisor_ollama_test.go bare AdvisorContext/WithAdvisorContext/ToolCall
  references to coretool.* after wrapper deletion
